Hibernate removed, we use JdbcClient

+++ src/main/resources/db/migration/V0__create_users_table.sql
@@ -1,7 +1,9 @@
 CREATE EXTENSION IF NOT EXISTS "uuid-ossp";
 
 CREATE TABLE IF NOT EXISTS users (
-    id SERIAL,
+    -- SERIAL = java's integer or Integer(32-bit), BIGSERIAL(64-bit) = java's Long or long
+    -- https://www.postgresql.org/docs/current/datatype-numeric.html
+    id BIGSERIAL,
     email TEXT NOT NULL,
     username TEXT NOT NULL,
     password TEXT NOT NULL,
